refactor(post_processing): split 2462-line file into mod/artwork_copy/state
Phase 3 of the refactor plan. Three-file split:

  post_processing/
    mod.rs          1380  POST_PROC_LOCK, do_file_op (spawn_blocking 1-4GB
                           cross-device copy discipline preserved),
                           SeriesImportCtx, load_series_import_ctx,
                           import_torrent (~785 lines), run_once,
                           advance_state_without_import
    artwork_copy.rs  629  copy_artwork (self-healing fan-out),
                           PosterOutcome + copy_series_and_season_poster,
                           BannerOutcome + copy_series_banner_and_backdrop,
                           + 5 async tests (artwork fan-out pinning)
    state.rs         500  grab_is_stale, fallback_ep_offset,
                           LibraryClassifyReport / PendingClassification,
                           scan_library_for_unclassified,
                           scan_series_for_unclassified, scan_for_unclassified,
                           + 5 fallback_offset tests

Visibility bumps applied:
  - 3 artwork helpers → pub(super)
  - fallback_ep_offset → pub(super)
  - PosterOutcome / BannerOutcome structs + fields → pub(super) / pub
    (crossed module boundary when called from mod.rs's import path)

POST_PROC_LOCK stays in mod.rs per the plan's "singleton static path
mustn't drift" note; state.rs imports it via `use super::POST_PROC_LOCK;`
for the unclassified-scan code paths.

External callers (handlers/library, main.rs) still resolve:
  post_processing::grab_is_stale, ::import_torrent, ::run_once,
  ::run_once_inner, ::scan_library_for_unclassified,
  ::scan_series_for_unclassified via pub use re-exports.

cargo fmt / clippy -D warnings / build / test all green (601 tests).
