+++ src/services/post_processing/mod.rs
@@ -5,10 +5,17 @@ use std::sync::LazyLock;
 use crate::AppState;
 use crate::models::log::LogCategory;
 use crate::models::{
-    artwork_cache, config, episode_tags, grabbed_torrents, local_metadata, metadata_cache, series,
+    config, episode_tags, grabbed_torrents, local_metadata, metadata_cache, series,
 };
 use crate::services::source::{self, SeriesContext};
-use crate::services::{artwork, logger, media, nfo};
+use crate::services::{logger, media, nfo};
+
+mod artwork_copy;
+mod state;
+
+use artwork_copy::{copy_series_and_season_poster, copy_series_banner_and_backdrop};
+use state::fallback_ep_offset;
+pub use state::{grab_is_stale, scan_library_for_unclassified, scan_series_for_unclassified};
 
 static POST_PROC_LOCK: LazyLock<tokio::sync::Mutex<()>> =
     LazyLock::new(|| tokio::sync::Mutex::new(()));
@@ -21,19 +28,6 @@ static POST_PROC_LOCK: LazyLock<tokio::sync::Mutex<()>> =
 // client-agnostic checks. Pre-refactor this function only knew qBit's
 // string set, which silently skipped Deluge's completed torrents
 // forever (#63 Phase 2 regression).
-
-/// Check if a grab is older than `max_age_secs` seconds.
-pub fn grab_is_stale(grabbed_at: &str, max_age_secs: i64) -> bool {
-    // grabbed_at is SQLite CURRENT_TIMESTAMP format: "YYYY-MM-DD HH:MM:SS"
-    let Some(grab_time) =
-        chrono::NaiveDateTime::parse_from_str(grabbed_at, "%Y-%m-%d %H:%M:%S").ok()
-    else {
-        return false;
-    };
-    let now = chrono::Utc::now().naive_utc();
-    let elapsed = now.signed_duration_since(grab_time).num_seconds();
-    elapsed > max_age_secs
-}
 
 fn is_video_file(name: &str) -> bool {
     let lower = name.to_lowercase();
@@ -115,354 +109,6 @@ async fn do_file_op(mode: &str, src: &Path, dst: &Path) -> std::io::Result<()> {
     .map_err(|e| std::io::Error::other(format!("join error: {}", e)))?
 }
 
-/// Copy a cached artwork blob (poster or banner) to every path in
-/// `dests`, self-healing the cache on a miss when `source_url` is
-/// available. Returns a `Vec<bool>` aligned to `dests` — `true` at
-/// position `i` means the blob landed at `dests[i]`.
-///
-/// **Self-healing rationale:**
-/// `metadata_sync::refresh_series_metadata_inner` is the only path
-/// that eagerly caches series artwork, and every `artwork::cache_image`
-/// error there is swallowed via `let _ = ...`. A transient fetch
-/// failure (AL CDN hiccup, a cancelled rebuild sweep, Jikan-fallback
-/// with an empty banner field, …) leaves the blob missing
-/// indefinitely. Baking the re-fetch into `copy_artwork` itself
-/// (rather than a separate pre-step the caller has to remember to
-/// run) makes the self-heal an invariant: any future caller that
-/// supplies a source URL automatically inherits it.
-///
-/// **I/O shape:** the blob bytes are read into memory once and
-/// written to every dest under a single `spawn_blocking`, so a
-/// network-backed media root pays one open per import rather than
-/// `dests.len()` opens. Failures are non-fatal (Jellyfin can still
-/// scrape its own artwork) but every failure path warns with enough
-/// detail to debug it — silent "no artwork in Jellyfin" was the
-/// symptom the caller-side self-heal was designed to fix.
-///
-/// **Args:**
-/// - `cache_key` — e.g. `series-{id}-cover` or `series-{id}-banner`
-/// - `image_kind` — "cover" or "banner" (matches
-///   [`artwork::cache_image`]'s `image_kind`; also used verbatim in
-///   log text)
-/// - `source_url` — optional upstream URL to re-cache from on a miss;
-///   pass `None` when the caller has nothing to fetch from (e.g. a
-///   series that ran on the pure Jikan fallback with no banner field)
-async fn copy_artwork(
-    db: &sqlx::SqlitePool,
-    series_id: i64,
-    cache_key: &str,
-    image_kind: &str,
-    source_url: Option<&str>,
-    dests: &[&Path],
-) -> Vec<bool> {
-    if dests.is_empty() {
-        return Vec::new();
-    }
-
-    // Step 1: cache miss + non-empty source URL → self-heal by
-    // re-fetching. Two SELECTs on the unhappy path, one on the happy
-    // path (the re-check after the fetch is necessary because the
-    // fetch may have partially succeeded — e.g. blob written but ref
-    // upsert failed).
-    //
-    // `metadata_sync::refresh_series_metadata_inner` is the only
-    // upstream caller of `cache_image`, and every error there is
-    // swallowed via `let _ = ...`. A transient fetch failure (AL CDN
-    // hiccup, a cancelled rebuild sweep, a Jikan-fallback with an
-    // empty banner field, …) would otherwise leave the blob missing
-    // indefinitely — baking the re-fetch in here makes the self-heal
-    // an invariant that any future caller supplying a source URL
-    // inherits automatically.
-    let mut cached = artwork_cache::get(db, cache_key).await.ok().flatten();
-    if cached.is_none()
-        && let Some(url) = source_url.filter(|u| !u.trim().is_empty())
-    {
-        tracing::debug!(
-            target: "ryokan::post_processing",
-            series_id,
-            cache_key,
-            image_kind,
-            "artwork cache miss; fetching on demand",
-        );
-        if let Err(err) =
-            artwork::cache_image(db, cache_key, "series", Some(series_id), image_kind, url).await
-        {
-            logger::warn(
-                db,
-                LogCategory::PostProcess,
-                &format!(
-                    "On-demand artwork fetch failed for {image_kind} on series_id={series_id}"
-                ),
-                &format!("source_url={url}, error={err}"),
-            )
-            .await;
-        }
-        cached = artwork_cache::get(db, cache_key).await.ok().flatten();
-    }
-
-    let Some(entry) = cached else {
-        tracing::debug!(
-            target: "ryokan::post_processing",
-            series_id,
-            cache_key,
-            image_kind,
-            "artwork blob missing from cache and no recoverable source; skipping copy",
-        );
-        return vec![false; dests.len()];
-    };
-
-    // Step 2: read the blob bytes once, fan out to each dest. Under
-    // one `spawn_blocking` so an NFS-backed media root doesn't
-    // serialize file-open round-trips from separate tokio tasks.
-    //
-    // Outcome is either `SourceReadFailed` (one error, affects every
-    // dest) or `PerDest` (dest-specific errors). Keeping these
-    // distinct lets the caller log the source-read failure exactly
-    // once instead of N times with misleading `dst=` paths.
-    enum CopyOutcome {
-        SourceReadFailed(std::io::Error),
-        PerDest(Vec<Result<(), std::io::Error>>),
-    }
-
-    let src = std::path::PathBuf::from(&entry.local_path);
-    let owned_dests: Vec<std::path::PathBuf> = dests.iter().map(|p| p.to_path_buf()).collect();
-    let src_display = src.display().to_string();
-    let copy_result = tokio::task::spawn_blocking(move || -> CopyOutcome {
-        let bytes = match std::fs::read(&src) {
-            Ok(b) => b,
-            Err(e) => return CopyOutcome::SourceReadFailed(e),
-        };
-        // First dest gets the real write; subsequent dests are
-        // hardlinked to it when possible so a multi-dest fan-out
-        // spends one blob's worth of bytes on disk regardless of
-        // `dests.len()`. Motivating case: series-root `banner.jpg`
-        // + `backdrop.jpg` (same blob, same directory) would
-        // otherwise cost ~500 MB of pure duplication in a
-        // 1000-series library, plus the same amount on every
-        // rsync/backup/snapshot.
-        //
-        // Fallback to `fs::write` if the hardlink fails — cross-fs
-        // (backdrop dir vs. series-root dir on different mounts),
-        // unusual FS (FAT32, some SMB mounts), or the first write
-        // errored and the source doesn't exist to link against. The
-        // fallback is behaviorally identical to the pre-dedupe code.
-        let per_dest: Vec<std::io::Result<()>> = owned_dests
-            .iter()
-            .enumerate()
-            .map(|(i, dst)| -> std::io::Result<()> {
-                if let Some(parent) = dst.parent() {
-                    std::fs::create_dir_all(parent)?;
-                }
-                if i == 0 {
-                    // Anchor: actual bytes on disk. Subsequent
-                    // hardlinks reference this inode.
-                    std::fs::write(dst, &bytes)
-                } else {
-                    // Hardlink wants a nonexistent dst; clean any
-                    // stale file first. `remove_file` error is
-                    // ignored because "already absent" is the
-                    // happy case.
-                    let _ = std::fs::remove_file(dst);
-                    let anchor = &owned_dests[0];
-                    std::fs::hard_link(anchor, dst)
-                        .or_else(|_hardlink_err| std::fs::write(dst, &bytes))
-                }
-            })
-            .collect();
-        CopyOutcome::PerDest(per_dest)
-    })
-    .await;
-
-    match copy_result {
-        Ok(CopyOutcome::SourceReadFailed(err)) => {
-            logger::warn(
-                db,
-                LogCategory::PostProcess,
-                &format!("Failed to read cached {image_kind} blob for series_id={series_id}"),
-                &format!("src={}, error={}", src_display, err),
-            )
-            .await;
-            vec![false; dests.len()]
-        }
-        Ok(CopyOutcome::PerDest(per_dest)) => {
-            let mut results = Vec::with_capacity(dests.len());
-            for (i, r) in per_dest.into_iter().enumerate() {
-                match r {
-                    Ok(()) => results.push(true),
-                    Err(err) => {
-                        logger::warn(
-                            db,
-                            LogCategory::PostProcess,
-                            &format!(
-                                "Failed to write series {image_kind} for series_id={series_id}"
-                            ),
-                            &format!(
-                                "src={}, dst={}, error={}",
-                                src_display,
-                                dests[i].display(),
-                                err
-                            ),
-                        )
-                        .await;
-                        results.push(false);
-                    }
-                }
-            }
-            results
-        }
-        Err(join_err) => {
-            logger::warn(
-                db,
-                LogCategory::PostProcess,
-                &format!("{image_kind} copy task panicked for series_id={series_id}"),
-                &format!("src={}, error={}", src_display, join_err),
-            )
-            .await;
-            vec![false; dests.len()]
-        }
-    }
-}
-
-/// Named result of a poster fan-out to the two slots Jellyfin reads:
-/// the series-root `poster.jpg` (series-level card) and the season
-/// folder's `folder.jpg` (season-card poster). Structural naming so
-/// the caller can't accidentally swap the two booleans via a slice
-/// reorder.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-struct PosterOutcome {
-    /// True if the series-root `poster.jpg` landed on disk.
-    series_root: bool,
-    /// True if the `Season NN/folder.jpg` landed on disk.
-    season_folder: bool,
-}
-
-/// Copy the cached series poster blob to both the series-root
-/// `poster.jpg` and the season folder's `folder.jpg` — the two files
-/// Jellyfin reads for the series card + season card posters. The
-/// blob is self-healed from `source_url` if the cache is empty and
-/// read into memory once, fanned out to both dests under a single
-/// `spawn_blocking` (see [`copy_artwork`]).
-async fn copy_series_and_season_poster(
-    db: &sqlx::SqlitePool,
-    series_id: i64,
-    source_url: Option<&str>,
-    series_poster_dest: &Path,
-    season_folder_dest: &Path,
-) -> PosterOutcome {
-    let cache_key = format!("series-{}-cover", series_id);
-    let results = copy_artwork(
-        db,
-        series_id,
-        &cache_key,
-        "cover",
-        source_url,
-        &[series_poster_dest, season_folder_dest],
-    )
-    .await;
-    // `copy_artwork`'s Vec<bool> contract is index-aligned to the
-    // `dests` slice — mapping those indices onto the named fields
-    // here (rather than at the caller) keeps the positional
-    // dependency contained to this 3-line function, where a swap
-    // would be immediately obvious to any reader.
-    PosterOutcome {
-        series_root: results.first().copied().unwrap_or(false),
-        season_folder: results.get(1).copied().unwrap_or(false),
-    }
-}
-
-/// Named result of a banner fan-out to the two Jellyfin image slots
-/// we fill from the AniList `bannerImage`. Structural naming so the
-/// caller can't accidentally swap the two booleans via slice reorder.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-struct BannerOutcome {
-    /// True if `{series_folder}/banner.jpg` landed — Jellyfin
-    /// `ImageType::Banner` (3). Only surfaces in "Banner" library
-    /// layouts; mostly legacy in the current Jellyfin web UI.
-    series_banner: bool,
-    /// True if `{series_folder}/backdrop.jpg` landed — Jellyfin
-    /// `ImageType::Backdrop` (2). This is the slot the series detail
-    /// page reads for the hero image behind the header, and what the
-    /// home screen uses as the featured background. AniList's
-    /// `bannerImage` is semantically a backdrop (wide hero, 1900×400
-    /// typical), not a thin Kodi-style banner, so we copy the same
-    /// blob into both slots and let Jellyfin pick per UI context.
-    series_backdrop: bool,
-}
-
-/// Copy the cached AniList banner blob to the two series-level image
-/// slots Jellyfin cares about: `banner.jpg` (legacy banner slot) and
-/// `backdrop.jpg` (modern hero/fanart slot). One `cache_image` fetch
-/// on a miss, one `fs::read` of the cached blob, two `fs::write`s
-/// under a single `spawn_blocking`.
-///
-/// Background: Jellyfin 10.x's default UI barely renders `ImageType::
-/// Banner` — the prominent wide image on the series detail page is
-/// the Backdrop. Copying the AL banner only to `banner.jpg` left
-/// Jellyfin showing nothing (or scraping TVDB/TMDB) for the backdrop
-/// slot. Writing both files keeps the data path unambiguous: same
-/// source blob, two on-disk filenames so auto-discovery finds the
-/// right one per slot.
-async fn copy_series_banner_and_backdrop(
-    db: &sqlx::SqlitePool,
-    series_id: i64,
-    source_url: Option<&str>,
-    banner_dest: &Path,
-    backdrop_dest: &Path,
-) -> BannerOutcome {
-    let cache_key = format!("series-{}-banner", series_id);
-    let results = copy_artwork(
-        db,
-        series_id,
-        &cache_key,
-        "banner",
-        source_url,
-        &[banner_dest, backdrop_dest],
-    )
-    .await;
-    // Same positional-contract containment pattern as
-    // `copy_series_and_season_poster` — keep the slice-index
-    // mapping local to this 3-line helper.
-    BannerOutcome {
-        series_banner: results.first().copied().unwrap_or(false),
-        series_backdrop: results.get(1).copied().unwrap_or(false),
-    }
-}
-
-/// #30 — Decide what offset to subtract from a parsed filename episode
-/// number when the file isn't covered by a Phase 2 auto-expand route
-/// row. Used by the legacy single-series path so absolute-numbered
-/// releases of sequel cours (`[SubsPlease] Jujutsu Kaisen - 56` for
-/// JJK S3 E9 with prior-cour total 47) land on the correct relative
-/// episode.
-///
-/// Rule: if the series has a non-zero `cumulative_prior_episodes` AND
-/// the parsed number exceeds that cumulative, treat the filename as
-/// absolute-numbered and subtract. Otherwise return 0 (relative
-/// numbering, file renames to its parsed number as-is).
-///
-/// Example with JJK S3 (cumulative = 47, own episodes = 12):
-///   - raw = 56 (SubsPlease absolute) → 56 > 47 → offset = 47 → ep = 9. ✓
-///   - raw = 9 (Erai-raws relative) → 9 ≤ 47 → offset = 0 → ep = 9. ✓
-///   - raw = 25 (stray S2 E1 file that got mis-grabbed) → 25 ≤ 47 →
-///     offset = 0, file lands as E25 of S3 (which doesn't exist,
-///     subsequent rename fails loudly — correct behavior).
-///
-/// First-season entries (cumulative = 0) short-circuit to offset = 0
-/// so the legacy behavior is preserved everywhere absolute numbering
-/// can't apply.
-fn fallback_ep_offset(raw_ep_num: i32, cumulative_prior_episodes: i32) -> i32 {
-    if cumulative_prior_episodes > 0 && raw_ep_num > cumulative_prior_episodes {
-        cumulative_prior_episodes
-    } else {
-        0
-    }
-}
-
-/// Per-series derived state used during an import. Built once per target
-/// series_id on demand inside [`import_torrent`] so a multi-series batch
-/// grab doesn't re-fetch the same rows or re-create the same directories
-/// for every file. The single-series case fills exactly one entry; a
-/// Phase 2 auto-expanded batch fills one entry per sibling touched.
 struct SeriesImportCtx {
     series: series::Series,
     folder_name: String,
@@ -1698,765 +1344,4 @@ async fn advance_state_without_import(state: &AppState) -> Result<(), ()> {
     }
 
     Ok(())
-}
-
-/// Summary of one run of [`scan_library_for_unclassified`]. Used by the
-/// handler to build a user-facing report and by logging to record the
-/// effect of the background task.
-#[derive(Debug, Default)]
-pub struct LibraryClassifyReport {
-    pub series_scanned: usize,
-    pub files_scanned: usize,
-    pub files_classified: usize,
-    pub files_needing_review: usize,
-}
-
-/// One file queued up by the lock-held enumeration phase, carried over
-/// into the unlocked classification phase of `scan_library_for_unclassified`.
-struct PendingClassification {
-    series_id: i64,
-    series_status: String,
-    series_season_year: Option<i32>,
-    series_end_year: Option<i32>,
-    series_root: std::path::PathBuf,
-    file_path: std::path::PathBuf,
-    episode_number: i32,
-    /// Sanitized on-disk filename. Used as a fallback title for L1
-    /// classification when `original_torrent_name` is empty (externally
-    /// imported files that Ryokan never grabbed).
-    title: String,
-    /// Original torrent name from `grabbed_torrents`, if a Ryokan grab
-    /// exists for this episode. Passed to `classify_post_download` as
-    /// the L1 input when non-empty: the post-processing import step
-    /// sanitizes release tags out of filenames (e.g. "[SubsPlease] Foo
-    /// (1080p) [Batch]" becomes "Foo - S01E05 - Title.mkv"), so
-    /// classifying against the on-disk name yields `rule=empty` and
-    /// the episode shows as UNKNOWN forever. Looking up the original
-    /// grab lets us classify against the unsanitized release name.
-    original_torrent_name: String,
-    /// True when an `episode_quality_tags` row already exists for this
-    /// (series, episode) pair. Determines whether the persist step goes
-    /// through `update_classification` (UPDATE) or `record_grab` +
-    /// `mark_completed` (INSERT upsert + state flip).
-    row_exists: bool,
-}
-
-/// Walk every tracked series and (re-)classify on-disk video files that
-/// don't yet have a confident classification row. This is the Phase 2
-/// "library scan path" — it catches files that were imported outside of
-/// Ryokan's own grab pipeline (pre-existing rips, manual drops, migrations
-/// from another PVR), AND it self-heals rows that the classifier first
-/// saw as `Unknown` (low-confidence filename-only result) now that the
-/// file is actually on disk and the full ffprobe/dir-walk pipeline can
-/// run against it.
-///
-/// Skips files whose tag row is:
-///  - already classified with a confident non-empty, non-"unknown" source,
-///  - flagged `needs_review = 1` (user should resolve via the review
-///    queue — we don't want to race against them),
-///  - flagged `manual_override = 1` (user pinned the classification).
-///
-/// **Locking:** the enumeration phase (config read, `scan_series_folder`,
-/// existing-tag lookup, disk-existence filter) holds `POST_PROC_LOCK` so
-/// the work list is a consistent snapshot that can't be invalidated by a
-/// parallel `run_once`. Once the list is built, the lock is released
-/// before we shell out to ffprobe via `classify_post_download` — those
-/// calls can take hundreds of ms per file and would otherwise block the
-/// 1-minute `process_completed_downloads` background loop for the full
-/// duration of a large scan. The DB writes in the second phase rely on
-/// SQLite's normal write serialization; the worst-case race (a real
-/// import landing between our classify and persist steps on the same
-/// episode) leaves a single row briefly stale and self-heals on the
-/// next scan.
-pub async fn scan_library_for_unclassified(state: &AppState) -> LibraryClassifyReport {
-    scan_for_unclassified(state, None).await
-}
-
-/// Issue #53: same enumeration + classify pipeline as
-/// [`scan_library_for_unclassified`] but scoped to a single series.
-/// Called from the import flow (`handlers/library/crud::add_series`) as a
-/// one-shot tokio::spawn so a freshly-imported series with pre-existing
-/// files on disk gets a classification pass within seconds instead of
-/// waiting up to six hours for the next periodic sweep.
-pub async fn scan_series_for_unclassified(
-    state: &AppState,
-    series_id: i64,
-) -> LibraryClassifyReport {
-    scan_for_unclassified(state, Some(series_id)).await
-}
-
-async fn scan_for_unclassified(
-    state: &AppState,
-    only_series_id: Option<i64>,
-) -> LibraryClassifyReport {
-    let mut report = LibraryClassifyReport::default();
-
-    let pending: Vec<PendingClassification> = {
-        // Lock scope: enumerate only. Dropped before the classify loop.
-        let _guard = POST_PROC_LOCK.lock().await;
-
-        let cfg = match config::get_config(&state.db).await.ok().flatten() {
-            Some(c) => c,
-            None => return report,
-        };
-        if cfg.media_root.is_empty() {
-            return report;
-        }
-
-        let tracked = match only_series_id {
-            // Single-series fast path used by the import hook — skip the
-            // full library enumeration when we know which series to look
-            // at. Bail silently when the id doesn't resolve (deleted
-            // between spawn and run).
-            Some(id) => match series::get_by_id(&state.db, id).await.ok().flatten() {
-                Some(row) => vec![row],
-                None => return report,
-            },
-            None => series::get_all(&state.db).await.unwrap_or_default(),
-        };
-        let mut pending = Vec::new();
-
-        for row in &tracked {
-            if row.folder_name.is_empty() {
-                continue;
-            }
-            let disk_files = media::scan_series_folder(&cfg.media_root, &row.folder_name);
-            if disk_files.is_empty() {
-                continue;
-            }
-            report.series_scanned += 1;
-
-            let existing = episode_tags::get_for_series(&state.db, row.id)
-                .await
-                .unwrap_or_default();
-
-            // One bulk fetch of imported grabs for this series, used
-            // by every file in the inner loop below to look up the
-            // unsanitized torrent name. Replaces a pair of per-file
-            // queries (find_imported_for_episode + a fallback
-            // most_recent_imported_torrent_name_for_series) that ran
-            // inside the held POST_PROC_LOCK — that was ~4800
-            // round-trips per pass on a 100-series, 24-ep library.
-            let imported_grabs = grabbed_torrents::imported_grabs_for_series(&state.db, row.id)
-                .await
-                .unwrap_or_default();
-            // Per-episode lookup; first occurrence (DESC by grabbed_at)
-            // wins, so each episode maps to its most recent grab. This
-            // matches the prior find_imported_for_episode .next()
-            // semantics on a DESC-sorted result set.
-            let mut grab_name_for_episode: HashMap<i32, String> = HashMap::new();
-            for (name, eps) in &imported_grabs {
-                for ep in eps {
-                    grab_name_for_episode
-                        .entry(*ep)
-                        .or_insert_with(|| name.clone());
-                }
-            }
-            let most_recent_grab_name: Option<&str> =
-                imported_grabs.first().map(|(n, _)| n.as_str());
-
-            let series_root = Path::new(&cfg.media_root).join(&row.folder_name);
-
-            for file in &disk_files {
-                report.files_scanned += 1;
-
-                // Decide whether to (re-)classify this file.
-                //
-                // Skip when:
-                //  - `manual_override = 1`: the user explicitly pinned
-                //    this classification — never touch.
-                //  - a tag exists with a non-empty source that isn't
-                //    "unknown". Already confidently classified. If such
-                //    a row is *also* flagged `needs_review`, we respect
-                //    that pending user decision and leave it alone.
-                //
-                // Pick up when:
-                //  - no tag exists yet (externally-imported file).
-                //  - the source column is empty (pre-Phase-1b rows).
-                //  - the source is literally "unknown" (case-insensitive)
-                //    — the classifier couldn't decide at grab time but
-                //    the file exists now, so retry with the full
-                //    ffprobe/dir-walk pipeline. This is the background
-                //    self-healing path for rows that started as Unknown.
-                //
-                // Note: unknown rows are retried **even when
-                // `needs_review = 1`**. That used to trap stale bad
-                // classifications — a row that was classified
-                // against the sanitized post-import filename before
-                // the original-torrent-name lookup landed produces
-                // `rule=empty` and gets flagged for review, and the
-                // old guard then refused to retry it even after the
-                // bug was fixed. An unknown needs-review row carries
-                // no useful user decision to stomp on (the user
-                // resolves via `set_manual_override`, which sets
-                // `manual_override = 1` and clears `needs_review`),
-                // so retrying is safe.
-                let tag = existing.get(&file.episode_number);
-                if let Some(t) = tag {
-                    if t.manual_override {
-                        continue;
-                    }
-                    let src = t.source.trim();
-                    let is_unknown = src.is_empty() || src.eq_ignore_ascii_case("unknown");
-                    if !is_unknown {
-                        continue;
-                    }
-                    // Issue #53: an UNKNOWN row that was already attempted
-                    // by the full-pipeline classifier (ffprobe + dir +
-                    // group + temporal + filename) won't change verdict
-                    // on the same bytes. Skip it so we don't re-ffprobe
-                    // every six hours forever — the user can still force
-                    // a fresh attempt by clearing/re-applying a manual
-                    // override or running the sweep manually after
-                    // updating the source-pipeline rules.
-                    if t.classification_attempted_at.is_some() {
-                        continue;
-                    }
-                }
-
-                // Reconstruct the absolute path so ffprobe can read the file.
-                let file_path = series_root.join(&file.filename);
-                if !file_path.exists() {
-                    continue;
-                }
-
-                // Use the sanitized on-disk filename as the fallback
-                // title. For Ryokan-grabbed files we'll override this
-                // with the original torrent name in the unlocked
-                // classify phase below so L1 sees the full release tag
-                // set instead of the tags-stripped post-import name.
-                let title = Path::new(&file.filename)
-                    .file_name()
-                    .and_then(|n| n.to_str())
-                    .unwrap_or(&file.filename)
-                    .to_string();
-
-                // Look up the grab that covers this episode so we can
-                // classify against the unsanitized torrent name. The
-                // pre-built `grab_name_for_episode` map and
-                // `most_recent_grab_name` fallback come from the
-                // single bulk fetch above — both lookups are now
-                // pure-Rust and don't touch the DB inside the held
-                // POST_PROC_LOCK. The fallback exists for old batch
-                // grabs recorded with `episode_numbers = []` before
-                // the grab-time episode-range fix; for a single-series
-                // batch the most-recent grab is by definition the one
-                // these files came from.
-                let original_torrent_name = grab_name_for_episode
-                    .get(&file.episode_number)
-                    .cloned()
-                    .or_else(|| most_recent_grab_name.map(|s| s.to_string()))
-                    .unwrap_or_default();
-
-                pending.push(PendingClassification {
-                    series_id: row.id,
-                    series_status: row.status.clone(),
-                    series_season_year: row.season_year,
-                    series_end_year: row.end_year,
-                    series_root: series_root.clone(),
-                    file_path,
-                    episode_number: file.episode_number,
-                    title,
-                    original_torrent_name,
-                    row_exists: tag.is_some(),
-                });
-            }
-        }
-
-        pending
-    };
-
-    // Unlocked classification + persist phase. ffprobe shell-outs are the
-    // slow part; doing them here instead of under the lock lets
-    // `process_completed_downloads` keep running in parallel.
-    for item in pending {
-        // Prefer the original torrent name from `grabbed_torrents`
-        // (carries full release tags like "[SubsPlease] Foo (01-12)
-        // [BD 1080p]"), fall back to the sanitized on-disk filename
-        // for externally-imported files that Ryokan never grabbed.
-        // This is the difference between L1 producing useful
-        // evidence and producing `rule=empty` on a release that
-        // landed via post-processing's rename step.
-        let classify_title: &str = if !item.original_torrent_name.is_empty() {
-            &item.original_torrent_name
-        } else {
-            &item.title
-        };
-        let is_batch = crate::models::grabbed_torrents::get_is_batch_by_name(
-            &state.db,
-            item.series_id,
-            classify_title,
-        )
-        .await
-        .unwrap_or(false);
-        let result = source::classify_post_download(
-            &state.db,
-            &item.file_path,
-            Some(&item.series_root),
-            classify_title,
-            Some(SeriesContext {
-                status: &item.series_status,
-                season_year: item.series_season_year,
-                end_year: item.series_end_year,
-            }),
-            is_batch,
-        )
-        .await;
-
-        // The row may not exist yet for externally-imported files, so
-        // we can't rely on `update_classification` alone (it's an
-        // UPDATE, not an UPSERT). Use `record_grab` with synthetic
-        // release metadata to insert-or-upsert, then flip state to
-        // 'completed' since the file is already on disk.
-        if !item.row_exists {
-            // Best-effort single stat — failure just leaves size at 0.
-            let file_size = tokio::fs::metadata(&item.file_path)
-                .await
-                .map(|m| m.len() as i64)
-                .unwrap_or(0);
-            // Externally-imported file — we're creating both the quality
-            // tag and the grab history row from thin air. `is_batch` is
-            // looked up above from the optional matching grabbed_torrents
-            // row; falls back to false for truly external files.
-            let _ = episode_tags::record_grab(
-                &state.db,
-                item.series_id,
-                item.episode_number,
-                &result,
-                classify_title,
-                "",
-                file_size,
-                is_batch,
-            )
-            .await;
-            let _ = episode_tags::mark_completed(&state.db, item.series_id, &[item.episode_number])
-                .await;
-            // Issue #53: stamp classification_attempted_at so the next
-            // sweep skips this row if `result` came back UNKNOWN. The
-            // grab-time path of `record_grab` deliberately leaves the
-            // column NULL, so we set it explicitly here for the
-            // post-classify path.
-            let _ = episode_tags::stamp_classification_attempted(
-                &state.db,
-                item.series_id,
-                item.episode_number,
-            )
-            .await;
-            // Flip the fresh grab history row to 'completed' and stamp
-            // in the on-disk file basename for the episode detail modal.
-            let imported_basename = item
-                .file_path
-                .file_name()
-                .and_then(|n| n.to_str())
-                .unwrap_or(classify_title)
-                .to_string();
-            let _ = episode_tags::mark_grab_history_completed(
-                &state.db,
-                item.series_id,
-                item.episode_number,
-                &imported_basename,
-                file_size,
-            )
-            .await;
-        } else {
-            // `update_classification` already sets
-            // classification_attempted_at internally — no extra stamp
-            // needed on this branch.
-            let _ = episode_tags::update_classification(
-                &state.db,
-                item.series_id,
-                item.episode_number,
-                &result,
-            )
-            .await;
-        }
-
-        report.files_classified += 1;
-        if result.needs_review {
-            report.files_needing_review += 1;
-        }
-    }
-
-    logger::info(
-        &state.db,
-        LogCategory::PostProcess,
-        "Library classify scan finished",
-        &format!(
-            "series={}, files_scanned={}, classified={}, needs_review={}",
-            report.series_scanned,
-            report.files_scanned,
-            report.files_classified,
-            report.files_needing_review
-        ),
-    )
-    .await;
-
-    report
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // ── #30 fallback offset selection ─────────────────────────────────
-
-    #[test]
-    fn fallback_offset_absolute_jjk_s3_subsplease() {
-        // Motivating case: [SubsPlease] Jujutsu Kaisen - 56 → S3 E9.
-        // S1 (24) + S2 (23) = 47 prior-cour episodes. Parsed 56 > 47,
-        // so the filename is absolute-numbered and offset = 47.
-        assert_eq!(fallback_ep_offset(56, 47), 47);
-    }
-
-    #[test]
-    fn fallback_offset_relative_release_within_cour() {
-        // Erai-raws / cour-specific releases number from 1: raw 9 is
-        // relative, 9 ≤ 47, offset = 0. Subtracting zero leaves E9.
-        assert_eq!(fallback_ep_offset(9, 47), 0);
-    }
-
-    #[test]
-    fn fallback_offset_zero_for_first_season_entry() {
-        // First-season entry has no prior cours (cumulative = 0), so
-        // the legacy behavior is preserved regardless of parsed number.
-        assert_eq!(fallback_ep_offset(56, 0), 0);
-        assert_eq!(fallback_ep_offset(1, 0), 0);
-    }
-
-    #[test]
-    fn fallback_offset_mis_grabbed_prior_cour_number() {
-        // Pathological: parsed 25 (S2 E1) somehow downloaded into the
-        // S3 series folder. 25 ≤ 47 → offset = 0, file lands as E25 of
-        // S3. S3 doesn't have E25 so the subsequent rename step fails
-        // — the right loud failure rather than silently translating to
-        // a plausible-looking E-22.
-        assert_eq!(fallback_ep_offset(25, 47), 0);
-    }
-
-    #[test]
-    fn fallback_offset_equal_to_cumulative() {
-        // Boundary: raw exactly equals cumulative. That value would be
-        // ambiguous (last episode of the prior cour, or legitimate
-        // relative E47 of a show with 47+ episodes). Without more
-        // signal we keep it as relative (offset 0) — the alternative
-        // (offset = cumulative) would silently map legitimate E47
-        // releases of a 48-episode show to E0.
-        assert_eq!(fallback_ep_offset(47, 47), 0);
-    }
-
-    // ── copy_artwork fan-out ──────────────────────────────────────────────
-    //
-    // The multi-dest fan-out in `copy_artwork` is the headline of the
-    // PR-51 review's dedupe item — one `std::fs::read` under a single
-    // `spawn_blocking`, N `fs::write`s to each dest. A regression that
-    // collapses back to per-dest spawns (or reads the blob once per
-    // dest) would be invisible to the existing NFO-level tests, so we
-    // pin the contract directly here.
-
-    /// Create just the two artwork tables we need, without dragging in
-    /// the full `models::migrate` schema. FKs aren't enforced on
-    /// in-memory SQLite unless `PRAGMA foreign_keys = ON` is set, so
-    /// the missing `series` parent row is harmless.
-    async fn setup_artwork_only_db() -> sqlx::SqlitePool {
-        let db = sqlx::SqlitePool::connect("sqlite::memory:")
-            .await
-            .expect("open in-memory db");
-        sqlx::query(
-            r#"CREATE TABLE image_blobs (
-                blob_hash TEXT PRIMARY KEY,
-                local_path TEXT NOT NULL DEFAULT '',
-                content_type TEXT NOT NULL DEFAULT '',
-                byte_size INTEGER NOT NULL DEFAULT 0,
-                created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-            )"#,
-        )
-        .execute(&db)
-        .await
-        .expect("create image_blobs");
-        sqlx::query(
-            r#"CREATE TABLE image_refs (
-                cache_key TEXT PRIMARY KEY,
-                parent_kind TEXT NOT NULL DEFAULT '',
-                parent_id INTEGER,
-                image_kind TEXT NOT NULL DEFAULT '',
-                source_url TEXT NOT NULL DEFAULT '',
-                blob_hash TEXT NOT NULL,
-                last_write INTEGER NOT NULL DEFAULT 0,
-                cached_at DATETIME DEFAULT CURRENT_TIMESTAMP
-            )"#,
-        )
-        .execute(&db)
-        .await
-        .expect("create image_refs");
-        // The `logs` table is used by `logger::warn` on the failure
-        // paths; create a minimal shape so those calls don't spuriously
-        // fail the test when exercising the error branches.
-        sqlx::query(
-            r#"CREATE TABLE logs (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                level TEXT NOT NULL DEFAULT '',
-                category TEXT NOT NULL DEFAULT '',
-                message TEXT NOT NULL DEFAULT '',
-                detail TEXT NOT NULL DEFAULT ''
-            )"#,
-        )
-        .execute(&db)
-        .await
-        .expect("create logs");
-        db
-    }
-
-    fn unique_test_dir(label: &str) -> std::path::PathBuf {
-        let nonce = format!(
-            "ryokan_pp_test_{}_{}_{}",
-            std::process::id(),
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_nanos())
-                .unwrap_or(0),
-            label,
-        );
-        let dir = std::env::temp_dir().join(nonce);
-        std::fs::create_dir_all(&dir).expect("create tempdir");
-        dir
-    }
-
-    async fn register_blob(
-        db: &sqlx::SqlitePool,
-        cache_key: &str,
-        blob_path: &std::path::Path,
-        blob_hash: &str,
-        byte_size: i64,
-    ) {
-        artwork_cache::upsert_blob(
-            db,
-            blob_hash,
-            &blob_path.to_string_lossy(),
-            "image/jpeg",
-            byte_size,
-        )
-        .await
-        .expect("upsert_blob");
-        artwork_cache::upsert_ref(
-            db,
-            artwork_cache::RefUpsert {
-                cache_key,
-                parent_kind: "series",
-                parent_id: Some(42),
-                image_kind: "cover",
-                source_url: "",
-                blob_hash,
-                last_write: 0,
-            },
-        )
-        .await
-        .expect("upsert_ref");
-    }
-
-    #[tokio::test]
-    async fn copy_artwork_fans_out_single_blob_to_multiple_dests() {
-        let db = setup_artwork_only_db().await;
-        let dir = unique_test_dir("copy_fanout");
-        let blob_path = dir.join("blob.jpg");
-        let payload = b"\xFF\xD8\xFF\xE0test jpeg body".to_vec();
-        std::fs::write(&blob_path, &payload).expect("write blob");
-        register_blob(
-            &db,
-            "series-42-cover",
-            &blob_path,
-            "deadbeef",
-            payload.len() as i64,
-        )
-        .await;
-
-        let dst_a = dir.join("poster.jpg");
-        let dst_b = dir.join("season/folder.jpg");
-
-        let results =
-            copy_artwork(&db, 42, "series-42-cover", "cover", None, &[&dst_a, &dst_b]).await;
-
-        assert_eq!(results, vec![true, true], "both dests must report success");
-        assert_eq!(
-            std::fs::read(&dst_a).expect("read dst_a"),
-            payload,
-            "dst_a bytes must match source",
-        );
-        assert_eq!(
-            std::fs::read(&dst_b).expect("read dst_b"),
-            payload,
-            "dst_b bytes must match source (nested dir must be created)",
-        );
-
-        // Cleanup — best effort.
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    /// Dedupe invariant: the fan-out must hardlink subsequent dests
-    /// to the first rather than writing the blob bytes twice. A
-    /// regression that silently restores per-dest writes would burn
-    /// on-disk bytes for every import (most visibly on banner +
-    /// backdrop, which always share the same blob in the same dir).
-    ///
-    /// Unix-only because this assertion leans on
-    /// `std::os::unix::fs::MetadataExt::{ino, nlink}` — the cheapest
-    /// way to prove hardlink-ness without shelling out to `stat`.
-    /// **For Windows support:** the production code path
-    /// ([`copy_artwork`]'s `std::fs::hard_link` call) already works
-    /// cross-platform (Linux/macOS `link(2)`, Windows
-    /// `CreateHardLinkW`) — only this test's assertion is gated.
-    /// To cover Windows, add a mirror test gated on `#[cfg(windows)]`
-    /// using `std::os::windows::fs::MetadataExt::number_of_links()`
-    /// (expect ≥ 2 on the anchor), or check file-index identity via
-    /// `BY_HANDLE_FILE_INFORMATION` through the `windows` crate for
-    /// a direct inode-equivalent comparison. The cross-platform
-    /// `copy_artwork_overwrites_preexisting_dest_files` test below
-    /// already runs on every target and pins *behavior* (both dests
-    /// end up with the new payload) independent of mechanism.
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn copy_artwork_hardlinks_subsequent_dests_to_the_first() {
-        use std::os::unix::fs::MetadataExt;
-
-        let db = setup_artwork_only_db().await;
-        let dir = unique_test_dir("copy_hardlink");
-        let blob_path = dir.join("blob.jpg");
-        let payload = b"\xFF\xD8\xFF\xE0anchor-inode-test".to_vec();
-        std::fs::write(&blob_path, &payload).expect("write blob");
-        register_blob(
-            &db,
-            "series-42-banner",
-            &blob_path,
-            "deadbeef",
-            payload.len() as i64,
-        )
-        .await;
-
-        // Both dests live in the same directory (same fs guaranteed
-        // on tmpfs/ext4) so hardlink must succeed without falling
-        // back to `fs::write`.
-        let banner_dst = dir.join("banner.jpg");
-        let backdrop_dst = dir.join("backdrop.jpg");
-
-        let results = copy_artwork(
-            &db,
-            42,
-            "series-42-banner",
-            "banner",
-            None,
-            &[&banner_dst, &backdrop_dst],
-        )
-        .await;
-
-        assert_eq!(results, vec![true, true]);
-
-        let banner_meta = std::fs::metadata(&banner_dst).expect("stat banner");
-        let backdrop_meta = std::fs::metadata(&backdrop_dst).expect("stat backdrop");
-        assert_eq!(
-            banner_meta.ino(),
-            backdrop_meta.ino(),
-            "backdrop.jpg must be hardlinked to banner.jpg, not a separate copy",
-        );
-        // Hardlinks share nlink count ≥ 2 (the anchor + at least one link).
-        assert!(
-            banner_meta.nlink() >= 2,
-            "banner.jpg nlink must reflect the hardlinked backdrop; got {}",
-            banner_meta.nlink(),
-        );
-
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    /// Hardlinking is best-effort: if the dst already exists (e.g.
-    /// a previous run landed it), the fan-out must clean it up
-    /// first. A regression that left the stale dst in place would
-    /// silently keep an old blob on one of the Jellyfin image
-    /// slots after an artwork refresh.
-    #[tokio::test]
-    async fn copy_artwork_overwrites_preexisting_dest_files() {
-        let db = setup_artwork_only_db().await;
-        let dir = unique_test_dir("copy_overwrite");
-        let blob_path = dir.join("blob.jpg");
-        let new_payload = b"new artwork blob".to_vec();
-        std::fs::write(&blob_path, &new_payload).expect("write blob");
-        register_blob(
-            &db,
-            "series-42-banner",
-            &blob_path,
-            "deadbeef",
-            new_payload.len() as i64,
-        )
-        .await;
-
-        let banner_dst = dir.join("banner.jpg");
-        let backdrop_dst = dir.join("backdrop.jpg");
-        // Pre-seed both dests with stale bytes.
-        std::fs::write(&banner_dst, b"stale banner").expect("seed banner");
-        std::fs::write(&backdrop_dst, b"stale backdrop").expect("seed backdrop");
-
-        let results = copy_artwork(
-            &db,
-            42,
-            "series-42-banner",
-            "banner",
-            None,
-            &[&banner_dst, &backdrop_dst],
-        )
-        .await;
-
-        assert_eq!(results, vec![true, true]);
-        assert_eq!(std::fs::read(&banner_dst).unwrap(), new_payload);
-        assert_eq!(std::fs::read(&backdrop_dst).unwrap(), new_payload);
-
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn copy_artwork_returns_all_false_when_cache_entry_missing_and_no_source() {
-        // Neither a cache row nor a source URL — function must degrade
-        // gracefully to `[false; N]` rather than erroring upward.
-        let db = setup_artwork_only_db().await;
-        let dir = unique_test_dir("copy_nocache");
-        let dst_a = dir.join("poster.jpg");
-        let dst_b = dir.join("folder.jpg");
-
-        let results =
-            copy_artwork(&db, 42, "series-42-cover", "cover", None, &[&dst_a, &dst_b]).await;
-
-        assert_eq!(results, vec![false, false]);
-        assert!(!dst_a.exists(), "dst_a must not exist on cache miss");
-        assert!(!dst_b.exists(), "dst_b must not exist on cache miss");
-
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn copy_artwork_source_read_failure_fans_false_to_all_dests() {
-        // Cache row exists but the blob file on disk has been removed
-        // (e.g. manual cleanup of the blob cache dir). The source-read
-        // failure path must return `[false; N]` and log once at the
-        // call site rather than N times — exercised here by calling
-        // the function and confirming all dests report false and
-        // nothing lands on disk.
-        let db = setup_artwork_only_db().await;
-        let dir = unique_test_dir("copy_srcgone");
-        let blob_path = dir.join("blob.jpg");
-        // Register the ref as if the blob were real, then remove the
-        // file so `fs::read` fails.
-        std::fs::write(&blob_path, b"stub").expect("write stub");
-        register_blob(&db, "series-42-cover", &blob_path, "deadbeef", 4).await;
-        std::fs::remove_file(&blob_path).expect("unlink blob");
-
-        let dst_a = dir.join("poster.jpg");
-        let dst_b = dir.join("folder.jpg");
-
-        let results =
-            copy_artwork(&db, 42, "series-42-cover", "cover", None, &[&dst_a, &dst_b]).await;
-
-        assert_eq!(results, vec![false, false]);
-        assert!(!dst_a.exists());
-        assert!(!dst_b.exists());
-
-        std::fs::remove_dir_all(&dir).ok();
-    }
 }
